@@ -33,7 +33,7 @@ class SlackNotificationAgentSpec extends Specification {
   def "sends correct message for #status status"() {
     given:
     def actualMessage = new BlockingVariable<SlackAttachment>()
-    slack.sendMessage(*_) >> { message, channel, asUser ->
+    slack.sendMessage(*_) >> { message, channel_local, asUser ->
       actualMessage.set(message)
     }
 
@@ -59,8 +59,8 @@ class SlackNotificationAgentSpec extends Specification {
   def "appends custom message to #status message if present"() {
     given:
     def actualMessage = new BlockingVariable<SlackAttachment>()
-    slack.sendMessage(*_) >> { message, channel, asUser ->
-      actualMessage.set(message)
+    slack.sendMessage(*_) >> { message_local, channel_local, asUser ->
+      actualMessage.set(message_local)
     }
 
     when:
@@ -88,7 +88,7 @@ class SlackNotificationAgentSpec extends Specification {
   def "sends entirely custom message if customMessage field is present, performing text replacement if needed"() {
     given:
     def actualMessage = new BlockingVariable<SlackAttachment>()
-    slack.sendMessage(*_) >> { message, channel, asUser ->
+    slack.sendMessage(*_) >> { message, channel_local, asUser ->
       actualMessage.set(message)
     }
 
